src/subject_service/requirements.txt
Flask[async]==3.0.0
Flask-Cors==4.0.0
pydantic>=2.0.0
typing
requests
python-dotenv
pyopenssl